#75 merge api abstraction and service logic
