+++ app/src/main/ui/src/services/Users.ts
@@ -1,6 +1,13 @@
 import {createPaginationData, defaultPaginationData, PagedModel} from "../lib/pagination";
-import {UsersAPI} from "../api/UsersAPI";
 import {QueryOptions} from "../hooks/useAPI";
+import {Axios} from "../api/axios";
+
+export interface UsersResponse {
+    entries: User[];
+    start: number;
+    limit: number;
+    totalEntries: number;
+}
 
 export type User = {
     username: string;
@@ -15,13 +22,20 @@ export type UsersModel = PagedModel & {
 export const DefaultUsersModel: UsersModel = {users: [], pagination: defaultPaginationData}
 
 export const UsersService = {
-    async get(signal?: AbortSignal, opts?: QueryOptions): Promise<UsersModel> {
+    async get(signal?: AbortSignal,opts?: QueryOptions): Promise<UsersModel> {
         return new Promise<UsersModel>(async (resolve, reject) => {
             try {
-                const usersData = await UsersAPI.get(signal, opts);
+                const usersResponse = await Axios.get<UsersResponse>("/users", {
+                    params: opts,
+                    signal: signal
+                });
+                if(usersResponse.status < 200 || usersResponse.status > 299) {
+                    reject(new Error("failed to load user data: " + usersResponse.status))
+                }
+                const usersData = usersResponse.data;
                 const paginationModel = createPaginationData(usersData.start, usersData.limit, usersData.totalEntries)
                 let model: UsersModel = {users: usersData.entries, pagination: paginationModel}
-                resolve(model);
+                resolve(model)
             } catch (e) {
                 reject(e);
             }
@@ -30,7 +44,11 @@ export const UsersService = {
     async delete(userName: string): Promise<void> {
         return new Promise<void>(async (resolve, reject) => {
             try {
-                resolve(UsersAPI.delete(userName));
+                const usersResponse = await Axios.delete<UsersResponse>(`/users/${userName}`);
+                if(usersResponse.status !== 204) {
+                    reject(new Error("failed to delete user: " + usersResponse.status))
+                }
+                resolve()
             } catch (e) {
                 reject(e);
             }
